+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,6 @@
         android:theme="@style/Theme.ArganicMolecule2"
         tools:targetApi="31">
         <activity
-            android:name=".A8.Testing_Image_Display"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".A8.DB_stickerMessage_activity"
             android:exported="false">
             <meta-data
