+++ app/src/main/java/com/examples/arganicmolecule2/A8/DB_stickerMessage_activity.java
@@ -1,16 +1,79 @@
 package com.examples.arganicmolecule2.A8;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
+import android.content.Context;
 import android.os.Bundle;
 
 import com.examples.arganicmolecule2.R;
+import com.examples.arganicmolecule2.model.sticker;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
+import com.google.firebase.database.ValueEventListener;
+
+import java.util.ArrayList;
 
 public class DB_stickerMessage_activity extends AppCompatActivity {
+    private RecyclerView stickers;
+    private ImageAdapter customerAdapter;
+    ArrayList<sticker> stickerList;
+    Context context;
+    FirebaseDatabase firebaseDatabase;
+    DatabaseReference databaseReference;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_db_sticker_message);
+
+        stickers = findViewById(R.id.sticker);
+        stickers.setHasFixedSize(true);
+        stickers.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.HORIZONTAL,false));
+        firebaseDatabase = FirebaseDatabase.getInstance();
+        databaseReference = firebaseDatabase.getReference();
+        stickerList = new ArrayList<>();
+        ClearAll();
+        GetDataFromFirebase();
+    }
+
+    private void GetDataFromFirebase() {
+        Query query = databaseReference.child("Data");
+        query.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                ClearAll();
+                for(DataSnapshot datasnapshot: snapshot.getChildren()){
+                    sticker sticker = new sticker();
+                    sticker.setImage(datasnapshot.child("image").getValue().toString());
+                    sticker.setName(datasnapshot.child("name").getValue().toString());
+                    stickerList.add(sticker);
+                }
+                System.out.println(stickerList);
+                context = getApplicationContext();
+                customerAdapter = new ImageAdapter(stickerList, context);
+                stickers.setAdapter(customerAdapter);
+                customerAdapter.notifyDataSetChanged();
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+            }
+        });
+    }
+
+    private void ClearAll(){
+        if(stickerList!=null){
+            stickerList.clear();
+            if(customerAdapter!=null){
+                customerAdapter.notifyDataSetChanged();
+            }
+        }
+        stickerList = new ArrayList<>();
     }
 }
